// src/api/bindings/Expense.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Expense = { id: string, categoryId: string, categoryName: string, categoryColor: string | null, amount: number, expenseDate: string, method: string, payee: string | null, description: string | null, createdByName: string | null, createdAt: string, };